% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepR.R
\name{gep_load_dll}
\alias{gep_load_dll}
\title{Load gepR.dll}
\usage{
gep_load_dll(dll_file = "gepR.dll")
}
\arguments{
\item{dll_file}{the path and name of the dll file to be loaded.}
}
\description{
Load gepR.dll
}
