% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepR.R
\name{gep_score}
\alias{gep_score}
\title{Predict using a GEP model}
\usage{
gep_score(x, sol_file = "gep_sol.dat")
}
\arguments{
\item{x}{data frame containing the input data. Must have the same number of columns as in the training data}

\item{sol_file}{character string of the file name where the trained model is stored. Default = "gep_sol.dat".}
}
\value{
numeric vector containing the predicted values
}
\description{
Predict using a GEP model
}
