% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gepR.R
\name{gep_unload_dll}
\alias{gep_unload_dll}
\title{Unload gepR.dll}
\usage{
gep_unload_dll(dll_file = "gepR.dll")
}
\arguments{
\item{dll_file}{the path and name of the dll file to be unloaded.}
}
\description{
Unload gepR.dll
}
